+++ internal/app/val/chapter_export.go
@@ -47,9 +47,9 @@ type UnitExport struct {
 	PageIndex int `json:"page_index"`
 
 	// XCoord 表示翻译单元的 X 坐标
-	XCoord int `json:"x_coord"`
+	XCoord float64 `json:"x_coord"`
 	// YCoord 表示翻译单元的 Y 坐标
-	YCoord int `json:"y_coord"`
+	YCoord float64 `json:"y_coord"`
 
 	// IsBubble 表示该单元是否是气泡框
 	IsBubble bool `json:"is_bubble"`
